Splits SIMP into its own file and fixes test
TO-5145
