+++ base/src/models/Plato_PenaltyModel.cpp
@@ -50,45 +50,13 @@
  *
  */
 
-#include "Plato_PenaltyModel.hpp"
+#include "Plato_SIMP.hpp"
 #include "Plato_Parser.hpp"
-#include <math.h>
 
-#include <boost/archive/xml_oarchive.hpp>
-#include <boost/archive/xml_iarchive.hpp>
-BOOST_CLASS_EXPORT_IMPLEMENT(Plato::SIMP)
+#include <string>
 
-namespace Plato {
-
-/******************************************************************************/
-SIMP::SIMP(const Plato::InputData& input)
-/******************************************************************************/
+namespace Plato
 {
-  m_penaltyExponent = Get::Double(input,"PenaltyExponent");
-  m_minimumValue    = Get::Double(input,"MinimumValue");
-}
-
-/******************************************************************************/
-double SIMP::eval(double x)
-/******************************************************************************/
-{
-  if (x != 0.0)
-    return m_minimumValue+(1.0-m_minimumValue)*pow(x,m_penaltyExponent);
-  else
-    return m_minimumValue;
-}
-
-/******************************************************************************/
-double SIMP::grad(double x)
-/******************************************************************************/
-{
-  if (x != 0.0)
-    return m_minimumValue+
-           (1.0-m_minimumValue)*m_penaltyExponent*pow(x,m_penaltyExponent-1.0);
-  else
-    return m_minimumValue;
-
-}
 
 /******************************************************************************/
 PenaltyModel* PenaltyModelFactory::create(Plato::InputData& aInputData)
@@ -105,7 +73,6 @@ PenaltyModel* PenaltyModelFactory::create(Plato::InputData& aInputData)
   }
 
   return pModel;
-
 }
 
 } /* namespace Plato */
