+++ base/src/operations/unittest/Plato_Test_Operations_Boost.cpp
@@ -64,7 +64,7 @@
 #include "Plato_EnforceBounds.hpp"
 #include "Plato_UpdateProblem.hpp"
 #include "Plato_ComputeVolume.hpp"
-#include "Plato_PenaltyModel.hpp"
+#include "Plato_SIMP.hpp"
 #include "Plato_CSMMeshOutput.hpp"
 #include "Plato_SetUpperBounds.hpp"
 #include "Plato_SetLowerBounds.hpp"
